# doc/openapi.yaml
openapi: 3.0.0
info:
  description: Virtualization API
  license:
    name: HPE End User License Agreement
    url: https://www.hpe.com/us/en/software/licensing.html
  title: Virtualization API
  version: 1.2.0
servers:
  - url: https://us-west.api.greenlake.hpe.com
  - url: https://eu-west.api.greenlake.hpe.com
  - url: https://eu-central.api.greenlake.hpe.com
  - url: https://ap-northeast.api.greenlake.hpe.com
security:
  - bearer: []
...
paths:
  /virtualization/v1beta1/virtual-machines:
    post:
      description: >
        Deploys one or more virtual machines in HCI environment with specified
        template and storage provisioning policy.
      operationId: HCIDeployVM
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HCIDeployVM"
        required: true
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
          description: Accepted
          headers:
            Location:
              schema:
                type: string
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized request
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal / unexpected error
      summary: Deploy virtual machine
      tags:
        - virtual-machines
  /virtualization/v1beta1/virtual-machines/{vm-id}:
    delete:
      description: Delete a virtual machine
      operationId: VMDelete
      parameters:
        - in: path
          name: vm-id
          required: true
          schema:
            $ref: "#/components/schemas/VirtualMachine/properties/id"
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
          description: Accepted
          headers:
            Location:
              schema:
                type: string
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized request
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal / unexpected error
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Appliance in maintenance mode
      summary: Delete a virtual machine
      tags:
        - virtual-machines
  /virtualization/v1beta1/virtual-machines/{vm-id}/power-off:
    post:
      description: Power off a virtual machine
      operationId: VMPowerOff
      parameters:
        - in: path
          name: vm-id
          required: true
          schema:
            $ref: "#/components/schemas/VirtualMachine/properties/id"
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
          description: Accepted
          headers:
            Location:
              schema:
                type: string
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized request
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal / unexpected error
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Appliance in maintenance mode
      summary: Power off a virtual machine
      tags:
        - virtual-machines
  /virtualization/v1beta1/virtual-machines/{vm-id}/power-on:
    post:
      description: Power on a virtual machine
      operationId: VMPowerOn
      parameters:
        - in: path
          name: vm-id
          required: true
          schema:
            $ref: "#/components/schemas/VirtualMachine/properties/id"
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
          description: Accepted
          headers:
            Location:
              schema:
                type: string
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized request
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal / unexpected error
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Appliance in maintenance mode
      summary: Power on a virtual machine
      tags:
        - virtual-machines
  /virtualization/v1beta1/virtual-machines/{vm-id}/refresh:
    post:
      description: Updates the properties of the specified virtual machine instance
      operationId: VMRefresh
      parameters:
        - in: path
          name: vm-id
          required: true
          schema:
            $ref: "#/components/schemas/VirtualMachine/properties/id"
      responses:
        "202":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/EmptyResponse"
          description: Request accepted
          headers:
            Location:
              schema:
                type: string
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthenticated request
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized request
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: A virtual machine instance with the given ID was not found.
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: >-
            A refresh action cannot be initiated in the current state of the
            cloud account.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: An unexpected error occurred.
      summary: Refresh the specified virtual machine instance
      tags:
        - virtual-machines
  /virtualization/v1beta1/virtual-machines/{vm-id}/reset:
    post:
      description: Reset a virtual machine
      operationId: VMPowerReset
      parameters:
        - in: path
          name: vm-id
          required: true
          schema:
            $ref: "#/components/schemas/VirtualMachine/properties/id"
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
          description: Accepted
          headers:
            Location:
              schema:
                type: string
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized request
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal / unexpected error
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Appliance in maintenance mode
      summary: Reset a virtual machine
      tags:
        - virtual-machines
  /virtualization/v1beta1/virtual-machines/{vm-id}/restart-guest-os:
    post:
      description: Restart guest OS of a virtual machine
      operationId: VMRestartGuestOS
      parameters:
        - in: path
          name: vm-id
          required: true
          schema:
            $ref: "#/components/schemas/VirtualMachine/properties/id"
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
          description: Accepted
          headers:
            Location:
              schema:
                type: string
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized request
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal / unexpected error
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Appliance in maintenance mode
      summary: Restart guest OS of a virtual machine
      tags:
        - virtual-machines
  /virtualization/v1beta1/virtual-machines/{vm-id}/shutdown-guest-os:
    post:
      description: Shutdown guest OS of a virtual machine
      operationId: VMShutdownGuestOS
      parameters:
        - in: path
          name: vm-id
          required: true
          schema:
            $ref: "#/components/schemas/VirtualMachine/properties/id"
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
          description: Accepted
          headers:
            Location:
              schema:
                type: string
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized request
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal / unexpected error
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Appliance in maintenance mode
      summary: Shutdown guest OS of a virtual machine
      tags:
        - virtual-machines
...
  /virtualization/v1beta1/virtual-machines-request/{request-id}:
    get:
      description: Details of a virtual machine request
      operationId: GetVirtualMachineRequest
      parameters:
        - in: path
          name: request-id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VMRequestWithDeploy"
          description: Success
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized request
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Resource not found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal / unexpected error
      summary: Get a virtual machine request identified by {request-id}
      tags:
        - virtual-machine-request
...
components:
  schemas:
...
    EmptyResponse:
      type: object
    ErrorResponse:
      type: object
      properties:
        debugId:
          description: A unique identifier for the request
          example: f57dcca3345820eb579c9317ce36dd92
          type: string
        errorCode:
          description: A machine friendly identifier for the error response
          type: string
        httpStatusCode:
          description: The HTTP status code of the response
          type: integer
        message:
          description: A user-friendly error message
          example: An example error message
          type: string
      required:
        - httpStatusCode
        - message
        - errorCode
        - debugId
...
    VMDeployInstance:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
        vmId:
          type: string
        vmName:
          type: string
        vmStatus:
          type: string
        vmStateMessage:
          type: string
        completedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - requestId
        - vmName
        - vmStatus
    VMRequestWithDeploy:
      type: object
      properties:
        vm_request:
          $ref: "#/components/schemas/VMRequest"
        vm_deploy_list:
          type: array
          items:
            $ref: "#/components/schemas/VMDeployInstance"
      required:
        - vm_request
        - vm_deploy_list
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      description: >
        The Data Service Cloud Console API uses a JWT bearer token for
        authentication.

        An authentication token can be obtained from the HPE GreenLake console.
      scheme: bearer
      type: http